Split customise-cms generator.js into focused modules
The 1199-line generator.js has been broken into logically cohesive
modules (each under 300 lines) while preserving the `generatePagesYaml`
public interface:

- generator-helpers.js: shared types, FieldContext, composition helpers
- blocks.js: block-schema -> CMS field conversion
- field-builders.js: non-item collections (pages, categories, team, ...)
- item-builders.js: item collections (news, products, events, ...)
- collection-config.js: dispatcher, view config, generateCollectionConfig
- static-configs.js: singleton file configs (site, meta, alt-tags)
- components.js: hoist _componentName markers into top-level components
- generator.js: orchestrator + custom blocks collection + generatePagesYaml
