+++ scripts/customise-cms/static-configs.js
@@ -1,0 +1,183 @@
+/**
+ * Static singleton-file configs for the CMS.
+ *
+ * These aren't collections — they're single JSON files in the site's `_data`
+ * directory (homepage settings, site metadata, schema.org meta, alt tags).
+ * Each generator takes the resolved `dataPath` so the same function works for
+ * templates with or without a `src/` folder.
+ */
+
+import { createObjectListField } from "#scripts/customise-cms/fields.js";
+
+/**
+ * @typedef {import('./generator-helpers.js').CollectionConfig} CollectionConfig
+ */
+
+/**
+ * Generate homepage settings configuration
+ * @param {string} dataPath - Path to data directory
+ * @returns {CollectionConfig} Homepage settings configuration
+ */
+export const getHomepageConfig = (dataPath) => ({
+  name: "homepage",
+  label: "Homepage Settings",
+  type: "file",
+  path: `${dataPath}/homepage.json`,
+  fields: [
+    {
+      name: "show_products",
+      type: "boolean",
+      label: "Show Products Section",
+      default: true,
+    },
+    {
+      name: "show_menus",
+      type: "boolean",
+      label: "Show Menus Section",
+      default: true,
+    },
+    {
+      name: "show_news",
+      type: "boolean",
+      label: "Show News Section",
+      default: true,
+    },
+    {
+      name: "show_recurring_events",
+      type: "boolean",
+      label: "Show Recurring Events",
+      default: true,
+    },
+  ],
+});
+
+/**
+ * Generate site configuration
+ * @param {string} dataPath - Path to data directory
+ * @returns {CollectionConfig} Site configuration
+ */
+export const getSiteConfig = (dataPath) => ({
+  name: "site",
+  label: "Site Configuration",
+  type: "file",
+  path: `${dataPath}/site.json`,
+  fields: [
+    { name: "name", type: "string", label: "Site Name" },
+    { name: "url", type: "string", label: "Site URL" },
+    {
+      name: "opening_times",
+      label: "Opening Times",
+      type: "object",
+      list: true,
+      fields: [
+        { name: "day", type: "string", label: "Day", required: true },
+        { name: "hours", type: "string", label: "Hours", required: true },
+      ],
+    },
+    {
+      name: "socials",
+      label: "Social Media Links",
+      type: "object",
+      fields: [
+        { name: "Github", type: "string", label: "Github" },
+        { name: "Forgejo", type: "string", label: "Forgejo" },
+        { name: "Facebook", type: "string", label: "Facebook" },
+        { name: "Instagram", type: "string", label: "Instagram" },
+        { name: "TikTok", type: "string", label: "TikTok" },
+        { name: "Google", type: "string", label: "Google" },
+        { name: "WhatsApp", type: "string", label: "WhatsApp" },
+        { name: "RSS", type: "string", label: "RSS" },
+      ],
+    },
+    { name: "map_embed_src", type: "string", label: "Map Embed URL" },
+  ],
+});
+
+/**
+ * Generate meta configuration
+ * @param {string} dataPath - Path to data directory
+ * @returns {CollectionConfig} Meta configuration
+ */
+export const getMetaConfig = (dataPath) => ({
+  name: "meta",
+  label: "Meta Configuration",
+  type: "file",
+  path: `${dataPath}/meta.json`,
+  fields: [
+    {
+      name: "language",
+      type: "string",
+      label: "Language Code",
+      default: "en-GB",
+    },
+    {
+      name: "organization",
+      label: "Organization",
+      type: "object",
+      fields: [
+        {
+          name: "description",
+          type: "string",
+          label: "Organization Description",
+        },
+        { name: "legalName", type: "string", label: "Legal Name" },
+        { name: "foundingDate", type: "string", label: "Founding Date" },
+        createObjectListField("founders", "Founders", [
+          { name: "name", type: "string", label: "Name" },
+        ]),
+        {
+          name: "address",
+          label: "Address",
+          type: "object",
+          fields: [
+            { name: "streetAddress", type: "string", label: "Street Address" },
+            { name: "addressLocality", type: "string", label: "City" },
+            { name: "addressRegion", type: "string", label: "Region/State" },
+            { name: "postalCode", type: "string", label: "Postal Code" },
+            { name: "addressCountry", type: "string", label: "Country Code" },
+          ],
+        },
+        {
+          name: "contactPoint",
+          label: "Contact Points",
+          type: "object",
+          list: true,
+          fields: [
+            { name: "telephone", type: "string", label: "Telephone" },
+            { name: "contactType", type: "string", label: "Contact Type" },
+            { name: "areaServed", type: "string", label: "Area Served" },
+            {
+              name: "availableLanguage",
+              type: "string",
+              label: "Available Languages",
+              list: true,
+            },
+          ],
+        },
+      ],
+    },
+  ],
+});
+
+/**
+ * Generate alt tags configuration
+ * @param {string} dataPath - Path to data directory
+ * @returns {CollectionConfig} Alt tags configuration
+ */
+export const getAltTagsConfig = (dataPath) => ({
+  name: "alt-tags",
+  label: "Image Alt Tags",
+  type: "file",
+  path: `${dataPath}/alt-tags.json`,
+  fields: [
+    {
+      name: "images",
+      type: "object",
+      list: true,
+      fields: [
+        { name: "path", type: "image", label: "Image" },
+        { name: "alt", type: "string", label: "Alt Text" },
+      ],
+    },
+  ],
+});
